pstree: Ignore processes that disappear before reading cmdline
The worst case here is that the user has specific a PID on the
command-line, read_proc() fails to read cmdline for an irrelevant,
transient process and fails.

It is better to simply ignore processes that have disappeared.  That
way pstree can provide useful output in more cases.

An alternative is to read cmdline only for processes that the user is
interested in.  However, this probably increases the chances of an
interesting process exiting before the cmdline is read.

Signed-off-by: Martin Schwenke <martin@meltin.net>